Show an interactive surface widget

This is a convenience wrapper around surfwidget for quick, interactive inspection of a single surface. It returns an HTML widget that can be used in R Markdown documents or Shiny applications.

Usage

show_surface_widget(x, data = NULL, ...)

Arguments

x

A SurfaceGeometry, NeuroSurface, or related object supported by surfwidget.

data

Optional numeric vector of data values for each vertex when x is a SurfaceGeometry. Ignored if x already carries data (e.g., a NeuroSurface).

...

Additional arguments passed on to surfwidget.

Value

An htmlwidget object.

Examples

# \donttest{
fs <- load_fsaverage_std8("inflated")
#> loading /home/runner/work/_temp/Library/neurosurf/extdata/std.8_lh.inflated.asc
#> loading /home/runner/work/_temp/Library/neurosurf/extdata/std.8_rh.inflated.asc
show_surface_widget(fs$lh)

# }